2.6 JUnit

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Calculator.java
@@ -10,7 +10,7 @@ public class Calculator {
         return a - b;
     }
 
-    public int square(int a) {
-        return (int) Math.pow(a, 2);
+    public double square(double a) {
+        return Math.pow(a, 2);
     }
 }
